Update files that were copied from the twentytwentyone theme
