--- README copy.md
@@ -1,80 +1,0 @@
-# Twenty Twenty-One
-
-Contributors: wordpressdotorg
-Requires at least: 5.3
-Tested up to: 5.6
-Requires PHP: 5.6
-License: GPLv2 or later
-License URI: http://www.gnu.org/licenses/gpl-2.0.html
-
-## Description
-
-Twenty Twenty-One is a blank canvas for your ideas and it makes the block editor your best brush.
-
-With new block patterns, which allow you to create a beautiful layout in a matter of seconds, this theme’s soft colors and eye-catching — yet timeless — design will let your work shine.
-
-Take it for a spin! See how Twenty Twenty-One elevates your portfolio, business website, or personal blog.
-
-To contribute to TT1 development, please read the [contributor's guide](/CONTRIBUTING.md).
-
-## Installation
-
-1. Click the 'Code' button on this Github page and select 'Download ZIP'.
-2. In your admin panel, go to Appearance -> Themes and click the 'Add New' button.
-3. Click on 'Upload Theme', select the zip file, and click on 'Install Now'.
-4. Click on the 'Activate' button to use your new theme right away.
-
-## Changelog
-
-## Copyright
-
-Twenty Twenty-One WordPress Theme, 2020 WordPress.org
-Twenty Twenty-One is distributed under the terms of the GNU GPL.
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 2 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU General Public License for more details.
-
-
-Twenty Twenty-One is derived from Seedlet, (C) 2020 Automattic, Inc.
-
-Twenty Twenty-One is also based on:
-
-Twenty Nineteen. 2018-2020 WordPress.org
-Twenty Nineteen is distributed under the terms of the GNU GPL v2 or later.
-
-Twenty Seventeen. Copyright (C) 2016-2020 WordPress.org
-Twenty Seventeen is distributed under the terms of the GNU GPL v2 or later.
-
-Twenty Twenty. Copyright (C) 2020 WordPress.org
-Twenty Twenty is distributed under the terms of the GNU GPL v2 or later.
-
-Underscores https://underscores.me/, Copyright (C) 2012-2020 Automattic, Inc.
-Underscores is distributed under the terms of the GNU GPL v2 or later.
-
-Normalizing styles have been helped along thanks to the fine work of
-Nicolas Gallagher and Jonathan Neal https://necolas.github.io/normalize.css/
-
-Unless otherwise noted, the icons in this theme are derived from the WordPress
-Icons Library, licensed under the terms of the GNU GPL v2 or later.
-https://github.com/WordPress/gutenberg/tree/master/packages/icons
-
-This theme bundles the following third-party resources:
-
-Images
-Art by Berthe Morisot 1841-1895. Public domain.
-https://www.wikiart.org/en/berthe-morisot/playing-in-the-sand
-https://www.wikiart.org/en/berthe-morisot/roses-tremieres-hollyhocks-1884
-https://www.wikiart.org/en/berthe-morisot/young-woman-in-mauve
-https://www.wikiart.org/en/berthe-morisot/in-the-bois-de-boulogne
-https://www.wikiart.org/en/berthe-morisot/daffodils
-https://www.wikiart.org/en/berthe-morisot/self-portrait-1885
-https://www.wikiart.org/en/berthe-morisot/the-garden-at-bougival-1884
-https://www.wikiart.org/en/berthe-morisot/villa-with-orange-trees-nice
-https://www.wikiart.org/en/berthe-morisot/reading
