+++ index.php
@@ -1,16 +1,38 @@
 <?php
 /**
- * The main template file.
+ * The main template file
  *
- * @package Irving_2021
+ * This is the most generic template file in a WordPress theme
+ * and one of the two required files for a theme (the other being style.css).
+ * It is used to display a page when nothing more specific matches a query.
+ * E.g., it puts together the home page when no home.php file exists.
+ *
+ * @link https://developer.wordpress.org/themes/basics/template-hierarchy/
+ *
+ * @package WordPress
+ * @subpackage Twenty_Twenty_One
+ * @since 1.0.0
  */
 
-?>
+get_header();
 
-<!DOCTYPE html>
-<html <?php language_attributes(); ?> class="no-js no-svg">
-	<?php wp_head(); ?>
-	<body <?php body_class(); ?>>
-	<?php wp_footer(); ?>
-	</body>
-</html>
+if ( have_posts() ) {
+
+	// Load posts loop.
+	while ( have_posts() ) {
+		the_post();
+
+		get_template_part( 'template-parts/content/content', get_theme_mod( 'display_excerpt_or_full_post', 'excerpt' ) );
+	}
+
+	// Previous/next page navigation.
+	twenty_twenty_one_the_posts_navigation();
+
+} else {
+
+	// If no content, include the "No posts found" template.
+	get_template_part( 'template-parts/content/content-none' );
+
+}
+
+get_footer();
