--- template-parts/footer.json
@@ -1,6 +1,0 @@
-{
-  "name": "irving/text",
-  "config": {
-    "content": "template-parts/footer.json"
-  }
-}
