--- template-parts/header.json
@@ -1,6 +1,0 @@
-{
-  "name": "irving/text",
-  "config": {
-    "content": "template-parts/header.json"
-  }
-}
