+++ template-parts/site-footer.json
@@ -1,0 +1,6 @@
+{
+  "name": "irving/text",
+  "config": {
+    "content": "template-parts/site-footer.json"
+  }
+}
